packages/experiment-registry/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "exp-registry"
version = "0.1.0"
description = "Structured YAML experiment registry for ML research"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.scripts]
exp = "exp_registry.cli:app"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/exp_registry"]
